refactor: let controllers own their capabilities, not the platforms
Entity platforms were probing controllers by string and branching on bed
type to build entities. This moves that knowledge onto BedController,
extending the pattern motor_control_specs already established for covers.

Capability contracts

Declares 11 previously-undeclared capabilities on BedController
(bed_presence_sides, auto_enable_massage, the climate side tuples, the
sleep-number setting bounds, the foundation preset options) and removes
57 of 68 getattr(controller, "...", default) probes. 26 of those probes
named properties the base already declared, so the default was dead; the
other 18 were implicit contracts no type checker verified, where a typo
returned the default forever. Two defaults silently disagreed with the
base (memory_slot_count passed 4 against a base of 0) but were
unreachable, since the property always exists.

The 11 remaining getattr calls are in diagnostics.py/support_report.py,
where they are deliberate optional introspection over per-controller
privates and emit a key only when present.

Alongside the probes, coordinator.py re-validated each seek tuning value
with isinstance. No real controller can fail those; they existed to
sanitize bare MagicMocks in tests. conftest gains make_controller_mock(),
which seeds a mock from the base class's own property defaults, so mocks
stay faithful as the contract grows instead of answering every capability
with a truthy Mock.

Position sliders

Adds PositionNumberSpec and BedController.position_number_specs, moving
number.py's six-way bed_type chain (243 lines of near-identical
description building) into the controllers that own each layout. The base
reproduces the standard back/legs/head/feet angle layout; Keeson,
SleepysBox25, SleepStar, OkinCst, Kaidi and SleepNumber override it.

The base deliberately does not derive from motor_control_specs:
OkinUuidController returns a remote-dependent subset of motors, so
deriving would change which sliders that bed exposes.

Reverie's 60-degree head/back limit moves onto the controllers via a new
motor_max_angles hook, so cover.py no longer checks bed types either.

Equivalence was verified by replaying the old branching against the new
specs: 14 bed types x 3 motor counts for sliders, and all 55 bed types x
3 motor counts (565 motor specs) for cover travel limits, both identical.

Shared predicate

bed_type_has_position_feedback() replaces six copies of the same
"BEDS_WITH_POSITION_FEEDBACK or Keeson-is-ergomotion" test, one of which
was already a private helper in config_flow that three inline copies in
that same file did not use.

Service handlers

Moves _async_register_services and its closures out of __init__.py into
services.py as module-level functions taking hass from call.hass. The
service constants move with them and are re-exported from the package
root. __init__.py: 1294 -> 546 lines.

Behavior change

number.py now requires a live controller to create position sliders. The
old code built them from bed_type alone for most beds, so a bed that
disconnects between a successful setup connect and platform forwarding
loses its sliders until a reload. The old defensiveness was already
inconsistent (Kaidi and SleepStar required a controller) and cover.py has
always required one; this makes the platforms uniform.

number.py bed-type references: 25 -> 7. cover.py: 5 -> 2.
