+++ custom_components/adjustable_bed/beds/reverie_nightstand.py
@@ -20,6 +20,7 @@ from typing import TYPE_CHECKING, Any
 from bleak.exc import BleakError
 
 from ..const import (
+    REVERIE_BACK_MAX_ANGLE,
     REVERIE_NIGHTSTAND_FOOT_POSITION_UUID,
     REVERIE_NIGHTSTAND_FOOT_WAVE_UUID,
     REVERIE_NIGHTSTAND_HEAD_POSITION_UUID,
@@ -99,6 +100,11 @@ class ReverieNightstandController(BedController):
         returns the linear head UUID as the primary one.
         """
         return REVERIE_NIGHTSTAND_LINEAR_HEAD_UUID
+
+    @property
+    def motor_max_angles(self) -> dict[str, float]:
+        """Reverie frames stop the head/back axis short of the standard 68 degrees."""
+        return {"back": REVERIE_BACK_MAX_ANGLE, "head": REVERIE_BACK_MAX_ANGLE}
 
     # Capability properties
     @property
