+++ custom_components/adjustable_bed/beds/sleepys_box25.py
@@ -30,7 +30,13 @@ from typing import TYPE_CHECKING
 from bleak.exc import BleakError
 
 from ..const import NORDIC_UART_WRITE_CHAR_UUID
-from .base import BedController, MotorControlSpec
+from .base import (
+    POSITION_UNIT_PERCENT,
+    BedController,
+    MotorControlSpec,
+    PositionNumberSpec,
+    build_position_number_spec,
+)
 
 if TYPE_CHECKING:
     from bleak.backends.characteristic import BleakGATTCharacteristic
@@ -246,6 +252,14 @@ class SleepysBox25Controller(BedController):
     @property
     def supports_direct_position_control(self) -> bool:
         return True
+
+    @property
+    def position_number_specs(self) -> tuple[PositionNumberSpec, ...]:
+        """Expose head/feet as percentage sliders; BOX25 reports percent, not angles."""
+        return (
+            build_position_number_spec("head", max_value=100.0, unit=POSITION_UNIT_PERCENT),
+            build_position_number_spec("feet", max_value=100.0, unit=POSITION_UNIT_PERCENT),
+        )
 
     @property
     def motor_control_specs(self) -> tuple[MotorControlSpec, ...]:
